Coded real control thread: CNewzflowThread receives messages to add NZB files to the queue. Real startup/exit conditions for CNewzflow. _CrtSetDbgFlag used in conjunction with "#define new" per source file for memory leak detection. Fixed leaks. Real basic queue management (click first toolbar button to add new NZB). Toolbar + Status Icons from µTorrent. Cleanup of CPostProcessor (all handles are now closed). Toolbar (toolbar.bmp) and Status Icons (tstatus.bmp) can be loaded from .bmp files in application directory for skinning purposes. (TODO: move to %APPDATA%\Newzflow)

+++ sock.h
@@ -45,7 +45,7 @@ public:
 	~CSpeedMonitor();
 
 	void OnReceive(int bytes);
-	float Get();
+	int Get();
 
 protected:
 	void Trim();
